Refactored Search to Share Artwork Pagination
